+++ src/adapter/10/IdMapper.ts
@@ -62,7 +62,6 @@ export function createIdMapper(): IdMapper {
 	return { has, update, getVNode, hasId, createId, getId, remove };
 }
 
-// FIXME: Preact 10 specific
 export function getInstance(vnode: VNode): any {
 	// For components we use the instance to check refs, otherwise
 	// we'll use a dom node