+++ src/adapter/setup.ts
@@ -1,43 +1,17 @@
 import { DevtoolsHook } from "./hook";
 import { Options } from "preact";
-import { createAdapter } from "./adapter/adapter";
-import { createBridge } from "./bridge";
 import { createRenderer } from "./10/renderer";
 import { setupOptions } from "./10/options";
-import { parseFilters } from "./10/filter";
 
 export async function init(options: Options, getHook: () => DevtoolsHook) {
-	const bridge = createBridge(window);
 	const hook = getHook();
 	const renderer = createRenderer(hook);
-	const adapter = createAdapter(hook, renderer);
 
 	// Add options as early as possible, so that we don't miss the first commit
 	setupOptions(options as any, renderer);
 
 	// Devtools can take a while to set up
 	await waitForHook(getHook);
-
-	bridge.listen("initialized", renderer.flushInitial);
-	bridge.listen("highlight", adapter.highlight);
-	bridge.listen("update-prop", ev => {
-		adapter.update(ev.id, "props", ev.path, ev.value);
-	});
-	bridge.listen("update-state", ev => {
-		adapter.update(ev.id, "state", ev.path, ev.value);
-	});
-	bridge.listen("update-context", ev => {
-		adapter.update(ev.id, "context", ev.path, ev.value);
-	});
-	bridge.listen("update-filter", ev => renderer.applyFilters(parseFilters(ev)));
-	bridge.listen("force-update", ev => renderer.forceUpdate(ev));
-	bridge.listen("select", adapter.select);
-	bridge.listen("copy", adapter.copy);
-	bridge.listen("inspect", adapter.inspect);
-	bridge.listen("log", adapter.log);
-	bridge.listen("update", adapter.log);
-	bridge.listen("start-picker", adapter.startPickElement);
-	bridge.listen("stop-picker", adapter.stopPickElement);
 
 	return hook.attach(renderer);
 }