+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Usability improvements
 - Minor Design tweaks
 - Upgrade dependencies and to Preact 10.0.5
+- Extract Preact X specific renderer code from `Adapter`
 
 ## 0.0.2